% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carleAndStrub.R
\name{carleAndStrub}
\alias{carleAndStrub}
\title{a helper function used by deplete to calculate the Carle and Strub estimate
for a removal experiment}
\usage{
carleAndStrub(x, alpha = 1, beta = 1)
}
\arguments{
\item{x}{a numeric vector with counts for each pass, or a dataframe that
includes the columns listed in passes}

\item{alpha}{the alpha parameter for the Carle.and.Strub method}

\item{beta}{the beta parameter for the Carle.and.Strub method}
}
\value{
\item{N}{population size estimate} \item{N.se}{population size
estimate standard error} \item{p}{capture probability estimate}
\item{p.se}{capture probability estimate standard error}
\item{N.CI95L}{lower 95\% CI for population size} \item{N.CI95U}{upper 95\%
CI for population size}
}
\description{
A function used by the deplete function. Not intended for stand alone use.
}
\details{
not intended for stand alone use.
}
\author{
Martin Liermann
}
\keyword{deplete}
