% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdeplete-package.R
\docType{data}
\name{IMWexample}
\alias{IMWexample}
\title{Removal data from form the IMW project}
\format{
A data frame with 100 observations on the following 6 variables.
\describe{ \item{list("tagYear")}{a numeric vector} \item{list("River")}{a
factor with levels \code{Deep Creek} \code{East Twin} \code{West Twin}}
\item{list("RP")}{a numeric vector} \item{list("Pass")}{a numeric vector}
\item{list("Species")}{a factor with levels \code{CO} \code{RBT}}
\item{list("CountOfSpecies")}{a numeric vector} }
}
\source{
a better description of the source
}
\description{
Removal data collected in conjunction with the Intensively Monitored
Watershed project on East Twin, West Twin, and Deep Creek on the Olympic
Penisula in Western Washington.
}
\details{
more details here
}
\references{
add reference here
}
\keyword{datasets}
