% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantyniemi.R
\name{mantyniemi}
\alias{mantyniemi}
\title{a helper function used by deplete to calculate the Mantyniemi estimate for a
removal experiment}
\usage{
mantyniemi(x, MCMCsims = F)
}
\arguments{
\item{x}{a numeric vector with counts for each pass, or a dataframe that
includes the columns listed in passes}

\item{MCMCsims}{if true Monte Carlo Markov Chain samples are returned}
}
\value{
\item{N}{population size estimate} \item{N.se}{population size
estimate standard error} \item{p}{capture probability estimate}
\item{p.se}{capture probability estimate standard error}
\item{N.CI95L}{lower 95\% CI for population size} \item{N.CI95U}{upper 95\%
CI for population size} \item{MCMCsims}{MCMC simulations returned if
MCMCsims=T}
}
\description{
A function used by the deplete function. Not intended for stand alone use.
}
\details{
not intended for stand alone use.
}
\author{
Martin Liermann
}
\keyword{deplete}
