% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdeplete-package.R
\docType{package}
\name{Rdeplete-package}
\alias{Rdeplete-package}
\alias{Rdeplete}
\title{A package for calculating depletion estimates}
\description{
includes functions to estimate population size based on depletion and a
function to simulate depletion data.
}
\details{
\tabular{ll}{ Package: \tab Rdeplete\cr Type: \tab Package\cr Version: \tab
1.1\cr Date: \tab 2012-10-25\cr License: \tab GPL (>= 2)\cr } More here
later
}
\author{
Martin Liermann: Who to complain to <martin.liermann@noaa.gov>
}
\keyword{package}
